# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Software City API - transforms code repos into living pixel-art cities
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface LoadRepoRequest {
  /** GitHub repository URL or local path */
  repoUrl: string;
  /** Branch to analyze (default main) */
  branch?: string;
}

export interface LoadRepoResponse {
  success: boolean;
  repoName: string;
  fileCount: number;
  districtCount: number;
  analysisTime: number;
  message: string;
}

export type DistrictType = (typeof DistrictType)[keyof typeof DistrictType];

export const DistrictType = {
  source: "source",
  test: "test",
  config: "config",
  api: "api",
  database: "database",
  docs: "docs",
  assets: "assets",
  root: "root",
} as const;

export type BuildingFileType =
  (typeof BuildingFileType)[keyof typeof BuildingFileType];

export const BuildingFileType = {
  class: "class",
  function: "function",
  api: "api",
  database: "database",
  config: "config",
  test: "test",
  entry: "entry",
  unknown: "unknown",
} as const;

export type BuildingStatus =
  (typeof BuildingStatus)[keyof typeof BuildingStatus];

export const BuildingStatus = {
  healthy: "healthy",
  warning: "warning",
  error: "error",
  fire: "fire",
  dark: "dark",
  glowing: "glowing",
} as const;

export type BuildingAge = (typeof BuildingAge)[keyof typeof BuildingAge];

export const BuildingAge = {
  new: "new",
  modern: "modern",
  aged: "aged",
  ancient: "ancient",
} as const;

export type BuildingActiveEvent =
  | (typeof BuildingActiveEvent)[keyof typeof BuildingActiveEvent]
  | null;

export const BuildingActiveEvent = {
  fire: "fire",
  flood: "flood",
  smoke: "smoke",
  alarm: "alarm",
  sparkle: "sparkle",
  collapse: "collapse",
} as const;

export interface Building {
  id: string;
  name: string;
  filePath: string;
  fileType: BuildingFileType;
  /** Height based on lines of code */
  floors: number;
  /** Cyclomatic complexity score */
  complexity: number;
  x: number;
  y: number;
  width: number;
  height: number;
  status: BuildingStatus;
  hasTests: boolean;
  testCoverage: number;
  commitCount: number;
  age: BuildingAge;
  language: string;
  linesOfCode: number;
  dependencies: string[];
  activeEvent?: BuildingActiveEvent;
}

export interface District {
  id: string;
  name: string;
  path: string;
  type: DistrictType;
  x: number;
  y: number;
  width: number;
  height: number;
  color: string;
  buildings: Building[];
  healthScore: number;
}

export type RoadType = (typeof RoadType)[keyof typeof RoadType];

export const RoadType = {
  import: "import",
  api: "api",
  database: "database",
  config: "config",
} as const;

export interface Road {
  id: string;
  fromBuilding: string;
  toBuilding: string;
  type: RoadType;
}

export type CityLayoutSeason =
  (typeof CityLayoutSeason)[keyof typeof CityLayoutSeason];

export const CityLayoutSeason = {
  summer: "summer",
  spring: "spring",
  autumn: "autumn",
  winter: "winter",
} as const;

export interface CityLayout {
  districts: District[];
  roads: Road[];
  repoName: string;
  totalFiles: number;
  season: CityLayoutSeason;
  healthScore: number;
  generatedAt: string;
}

export type CityHealthSeason =
  (typeof CityHealthSeason)[keyof typeof CityHealthSeason];

export const CityHealthSeason = {
  summer: "summer",
  spring: "spring",
  autumn: "autumn",
  winter: "winter",
} as const;

export type CityHealthBreakdown = {
  testCoverage: number;
  codeQuality: number;
  complexity: number;
  activity: number;
};

export interface CityHealth {
  /** 0-100 health score */
  score: number;
  season: CityHealthSeason;
  testCoverageRatio: number;
  cleanBuildingRatio: number;
  avgComplexity: number;
  commitFrequency: number;
  breakdown: CityHealthBreakdown;
}

export interface LiveMetrics {
  timestamp: string;
  cpuUsage: number;
  memoryUsage: number;
  activeAgents: number;
  bugsFound: number;
  testsRun: number;
  escalations: number;
  knowledgeBaseHits: number;
}

export type AgentRole = (typeof AgentRole)[keyof typeof AgentRole];

export const AgentRole = {
  qa_inspector: "qa_inspector",
  api_fuzzer: "api_fuzzer",
  load_tester: "load_tester",
  edge_explorer: "edge_explorer",
  ui_navigator: "ui_navigator",
} as const;

export type AgentStatus = (typeof AgentStatus)[keyof typeof AgentStatus];

export const AgentStatus = {
  idle: "idle",
  working: "working",
  escalating: "escalating",
  reporting: "reporting",
  waiting: "waiting",
} as const;

export interface Agent {
  id: string;
  name: string;
  role: AgentRole;
  status: AgentStatus;
  currentBuilding?: string | null;
  currentTask?: string | null;
  bugsFound: number;
  testsGenerated: number;
  escalations: number;
  accuracy: number;
  /** NPC level 1-5 based on accuracy */
  level: number;
  /** Current NPC status dialogue */
  dialogue: string;
  x: number;
  y: number;
  color: string;
}

export interface AgentList {
  agents: Agent[];
  total: number;
}

export type SpawnAgentRequestRole =
  (typeof SpawnAgentRequestRole)[keyof typeof SpawnAgentRequestRole];

export const SpawnAgentRequestRole = {
  qa_inspector: "qa_inspector",
  api_fuzzer: "api_fuzzer",
  load_tester: "load_tester",
  edge_explorer: "edge_explorer",
  ui_navigator: "ui_navigator",
} as const;

export interface SpawnAgentRequest {
  role: SpawnAgentRequestRole;
  targetBuilding?: string | null;
}

export type AgentTaskRequestTaskType =
  (typeof AgentTaskRequestTaskType)[keyof typeof AgentTaskRequestTaskType];

export const AgentTaskRequestTaskType = {
  generate_tests: "generate_tests",
  analyze_bug: "analyze_bug",
  fuzz_api: "fuzz_api",
  load_test: "load_test",
  explore_edge_cases: "explore_edge_cases",
} as const;

export interface AgentTaskRequest {
  taskType: AgentTaskRequestTaskType;
  buildingId: string;
  context: string;
}

export interface AgentTaskResult {
  success: boolean;
  taskType: string;
  result: string;
  actionItems: string[];
  bugsFound: number;
  escalated: boolean;
  fromKnowledgeBase: boolean;
  provider?: string | null;
}

export interface ChatRequest {
  message: string;
  buildingContext?: string | null;
}

export interface ChatResponse {
  agentId: string;
  agentName: string;
  message: string;
  escalated: boolean;
  fromKnowledgeBase: boolean;
  actionItems: string[];
}

export type KnowledgeStatsTopProblemsItem = {
  type: string;
  count: number;
};

export interface KnowledgeStats {
  totalEntries: number;
  totalCacheHits: number;
  avgBugsPerEntry: number;
  escalationRate: number;
  topProblems: KnowledgeStatsTopProblemsItem[];
}

export interface KnowledgeEntry {
  id: number;
  problemType: string;
  language: string;
  question: string;
  answer: string;
  confidence: string;
  provider: string;
  useCount: number;
  createdAt: string;
}

export interface KnowledgeEntryList {
  entries: KnowledgeEntry[];
  total: number;
}

export type CityEventType = (typeof CityEventType)[keyof typeof CityEventType];

export const CityEventType = {
  fire: "fire",
  bug_found: "bug_found",
  test_passed: "test_passed",
  escalation: "escalation",
  knowledge_hit: "knowledge_hit",
  building_collapse: "building_collapse",
  season_change: "season_change",
  agent_promoted: "agent_promoted",
} as const;

export type CityEventSeverity =
  (typeof CityEventSeverity)[keyof typeof CityEventSeverity];

export const CityEventSeverity = {
  info: "info",
  warning: "warning",
  error: "error",
  critical: "critical",
} as const;

export interface CityEvent {
  id: string;
  type: CityEventType;
  buildingId?: string | null;
  buildingName?: string | null;
  agentId?: string | null;
  agentName?: string | null;
  message: string;
  severity: CityEventSeverity;
  timestamp: string;
}

export interface CityEventList {
  events: CityEvent[];
  total: number;
}