# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Software City API - transforms code repos into living pixel-art cities
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /agents/spawn:
    post:
      operationId: spawnAgent
      tags: [agents]
      summary: Spawn a new NPC agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpawnAgentRequest"
      responses:
        "200":
          description: Agent spawned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
  /agents/{agentId}/task:
    post:
      operationId: assignAgentTask
      tags: [agents]
      summary: Assign a task to an NPC agent
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentTaskRequest"
      responses:
        "200":
          description: Task assigned and result returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentTaskResult"
  /agents/{agentId}/chat:
    post:
      operationId: chatWithAgent
      tags: [agents]
      summary: Chat with an NPC agent
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
      responses:
        "200":
          description: Agent response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
...
components:
  schemas:
...
    Agent:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [qa_inspector, api_fuzzer, load_tester, edge_explorer, ui_navigator]
        status:
          type: string
          enum: [idle, working, escalating, reporting, waiting]
        currentBuilding:
          type: string
          nullable: true
        currentTask:
          type: string
          nullable: true
        bugsFound:
          type: integer
        testsGenerated:
          type: integer
        escalations:
          type: integer
        accuracy:
          type: number
        level:
          type: integer
          description: NPC level 1-5 based on accuracy
        dialogue:
          type: string
          description: Current NPC status dialogue
        x:
          type: number
        y:
          type: number
        color:
          type: string
      required:
        - id
        - name
        - role
        - status
        - bugsFound
        - testsGenerated
        - escalations
        - accuracy
        - level
        - dialogue
        - x
        - y
        - color
...
    SpawnAgentRequest:
      type: object
      properties:
        role:
          type: string
          enum: [qa_inspector, api_fuzzer, load_tester, edge_explorer, ui_navigator]
        targetBuilding:
          type: string
          nullable: true
      required:
        - role
    AgentTaskRequest:
      type: object
      properties:
        taskType:
          type: string
          enum: [generate_tests, analyze_bug, fuzz_api, load_test, explore_edge_cases]
        buildingId:
          type: string
        context:
          type: string
      required:
        - taskType
        - buildingId
        - context
    AgentTaskResult:
      type: object
      properties:
        success:
          type: boolean
        taskType:
          type: string
        result:
          type: string
        actionItems:
          type: array
          items:
            type: string
        bugsFound:
          type: integer
        escalated:
          type: boolean
        fromKnowledgeBase:
          type: boolean
        provider:
          type: string
          nullable: true
      required:
        - success
        - taskType
        - result
        - actionItems
        - bugsFound
        - escalated
        - fromKnowledgeBase
    ChatRequest:
      type: object
      properties:
        message:
          type: string
        buildingContext:
          type: string
          nullable: true
      required:
        - message
    ChatResponse:
      type: object
      properties:
        agentId:
          type: string
        agentName:
          type: string
        message:
          type: string
        escalated:
          type: boolean
        fromKnowledgeBase:
          type: boolean
        actionItems:
          type: array
          items:
            type: string
      required:
        - agentId
        - agentName
        - message
        - escalated
        - fromKnowledgeBase
        - actionItems
...